lib: move i18n to agent/lib
Canonical home moves from scripts/lib to agent/lib; its settings.mjs
import goes back to a relative sibling path now that both live in
agent/lib, and every other importer picks up the matching alias.
Content is unchanged.
