+++ scripts/poller/deliver.mjs
@@ -1,6 +1,6 @@
 import { addTelegramQueueReceipt, TELEGRAM_ACCEPTANCE_KIND_HEADER } from "#lib/telegram-acceptance.mjs";
 import { classifyDeliverStatus } from "../lib/deliver-policy.mjs";
-import { tr } from "../lib/i18n.mjs";
+import { tr } from "#lib/i18n.mjs";
 import { ROUTE, ACCEPTANCE_ROUTE, SECRET, ALLOWED, SETTLE_MS, sleep, log } from "./config.mjs";
 import { tg } from "./transport.mjs";
 import { chatKey } from "./offset.mjs";
