+++ scripts/poller/queue.mjs
@@ -18,7 +18,7 @@ import {
   setChatStatus,
   setChatStatusIf,
 } from "#lib/run-status.mjs";
-import { tr } from "../lib/i18n.mjs";
+import { tr } from "#lib/i18n.mjs";
 import { DATA_DIR, SECRET, RESET_ROUTE, log } from "./config.mjs";
 import { tg } from "./transport.mjs";
