+++ scripts/poller/wizards.mjs
@@ -8,7 +8,7 @@ import {
 import { ModelValidationError, validateModelSelection } from "../lib/model-validation.mjs";
 import { getAccessToken, runDeviceCodeLogin } from "../lib/codex-oauth.mjs";
 import { compactNumber, modelSummary } from "../lib/model-summary.mjs";
-import { getLang, tr } from "../lib/i18n.mjs";
+import { getLang, tr } from "#lib/i18n.mjs";
 import { readEnvValues, upsertEnv } from "../lib/env-file.mjs";
 import { createFlows } from "../lib/tg-flow.mjs";
 import { ALLOWED, DATA_DIR_ABS, ENV_PATH, log } from "./config.mjs";
